update: 認証部分のロジックとビューを分離
